requirements.txt
vocode
fastapi
uvicorn
python-multipart
websockets
python-dotenv

